Introduce ForwardingBuildCacheService
+review REVIEW-6464
