+++ subprojects/core/src/main/java/org/gradle/caching/internal/BuildOperationFiringBuildCacheServiceDecorator.java
@@ -30,33 +30,22 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-public class BuildOperationFiringBuildCacheServiceDecorator implements BuildCacheService {
+public class BuildOperationFiringBuildCacheServiceDecorator extends ForwardingBuildCacheService {
     private final BuildOperationExecutor buildOperationExecutor;
-    private final BuildCacheService delegate;
 
     public BuildOperationFiringBuildCacheServiceDecorator(BuildOperationExecutor buildOperationExecutor, BuildCacheService delegate) {
+        super(delegate);
         this.buildOperationExecutor = buildOperationExecutor;
-        this.delegate = delegate;
     }
 
     @Override
     public boolean load(final BuildCacheKey key, final BuildCacheEntryReader reader) throws BuildCacheException {
-        return delegate.load(key, new BuildOperationFiringBuildCacheEntryReader(reader, key));
+        return super.load(key, new BuildOperationFiringBuildCacheEntryReader(reader, key));
     }
 
     @Override
     public void store(final BuildCacheKey key, final BuildCacheEntryWriter writer) throws BuildCacheException {
-        delegate.store(key, new BuildOperationFiringBuildCacheEntryWriter(writer, key));
-    }
-
-    @Override
-    public String getDescription() {
-        return delegate.getDescription();
-    }
-
-    @Override
-    public void close() throws IOException {
-        delegate.close();
+        super.store(key, new BuildOperationFiringBuildCacheEntryWriter(writer, key));
     }
 
     private class BuildOperationFiringBuildCacheEntryReader implements BuildCacheEntryReader {
