+++ subprojects/core/src/main/java/org/gradle/caching/internal/StagingBuildCacheServiceDecorator.java
@@ -18,11 +18,11 @@ package org.gradle.caching.internal;
 
 import org.apache.commons.io.IOUtils;
 import org.gradle.api.internal.file.TemporaryFileProvider;
-import org.gradle.caching.BuildCacheService;
 import org.gradle.caching.BuildCacheEntryReader;
 import org.gradle.caching.BuildCacheEntryWriter;
 import org.gradle.caching.BuildCacheException;
 import org.gradle.caching.BuildCacheKey;
+import org.gradle.caching.BuildCacheService;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
@@ -38,13 +38,12 @@ import java.io.OutputStream;
  * {@code BuildCacheService} decorator that stages files locally from a remote build cache. This provides a separation between
  * a build cache problem and a {@code BuildCacheEntryReader} or {@code BuildCacheEntryWriter} problem.
  */
-public class StagingBuildCacheServiceDecorator implements BuildCacheService {
-    private final BuildCacheService delegate;
+public class StagingBuildCacheServiceDecorator extends ForwardingBuildCacheService {
     private final boolean stageCacheEntries;
     private final TemporaryFileProvider temporaryFileProvider;
 
     public StagingBuildCacheServiceDecorator(TemporaryFileProvider temporaryFileProvider, boolean stageCacheEntries, BuildCacheService delegate) {
-        this.delegate = delegate;
+        super(delegate);
         this.stageCacheEntries = stageCacheEntries;
         this.temporaryFileProvider = temporaryFileProvider;
     }
@@ -56,29 +55,19 @@ public class StagingBuildCacheServiceDecorator implements BuildCacheService {
     @Override
     public boolean load(BuildCacheKey key, BuildCacheEntryReader reader) throws BuildCacheException {
         if (stageCacheEntries) {
-            return delegate.load(key, new StagingBuildCacheEntryReader(reader, temporaryFileProvider));
+            return super.load(key, new StagingBuildCacheEntryReader(reader, temporaryFileProvider));
         } else {
-            return delegate.load(key, reader);
+            return super.load(key, reader);
         }
     }
 
     @Override
     public void store(BuildCacheKey key, BuildCacheEntryWriter writer) throws BuildCacheException {
         if (stageCacheEntries) {
-            delegate.store(key, new StagingBuildCacheEntryWriter(writer, temporaryFileProvider));
+            super.store(key, new StagingBuildCacheEntryWriter(writer, temporaryFileProvider));
         } else {
-            delegate.store(key, writer);
+            super.store(key, writer);
         }
-    }
-
-    @Override
-    public String getDescription() {
-        return delegate.getDescription();
-    }
-
-    @Override
-    public void close() throws IOException {
-        delegate.close();
     }
 
     /**
